stop automated release every monday
sometimes we land syntax changes in yam repo; not very often but sometimes
for sure.
sometimes these changes requires changes in melange config inside
package repositories. Releasing in a automated fashion every is not
ideal because we use yam@latest in package repositories for checking.

ideally before cutting a tag; we will do a local build of yam and then
if it's breaking in b/w two versions then we'll coordinate the changes
slowly involving multiple repos.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>